[BE] refactor: error code를 모아서 관리하고 도메인별 exception을 분리한다.
