+++ BE/src/auth/exception/expired-token.exception.ts
@@ -1,0 +1,9 @@
+import { MotimateException } from '../../common/exception/motimate.excpetion';
+import { ERROR_INFO } from '../../common/exception/error-code';
+
+
+export class ExpiredTokenException extends MotimateException {
+  constructor() {
+    super(ERROR_INFO.EXPIRED_TOKEN);
+  }
+}
