+++ BE/src/common/exception/motimate.excpetion.ts
@@ -1,9 +1,5 @@
 import { HttpException } from '@nestjs/common';
-
-export interface ErrorCode {
-  statusCode: number;
-  message: string;
-}
+import { ErrorCode } from './error-code';
 
 export class MotimateException extends HttpException {
   constructor(errorCode: ErrorCode) {
